test(Effect): tighten timeoutOrElse coverage
